+++ mamba_ssm/ops/triton/ssd_bmm.py
@@ -50,7 +50,9 @@ def _bmm_chunk_fwd_kernel(
     BLOCK_SIZE_M: tl.constexpr, BLOCK_SIZE_N: tl.constexpr, BLOCK_SIZE_K: tl.constexpr,
 ):
     pid_b = tl.program_id(axis=1)
-    pid_ch = tl.program_id(axis=2)
+    # if chunk_size/stride products are large, may overflow int32, so use 64 bit
+    # https://github.com/triton-lang/triton/issues/1058
+    pid_ch = tl.program_id(axis=2).to(tl.int64)
     pid_c = pid_ch // ngroups
     pid_h = pid_ch - pid_c * ngroups
     num_pid_n = tl.cdiv(chunk_size, BLOCK_SIZE_N)
@@ -123,7 +125,9 @@ def _bmm_chunk_bwd_kernel(
     BLOCK_SIZE_M: tl.constexpr, BLOCK_SIZE_N: tl.constexpr, BLOCK_SIZE_CS: tl.constexpr,
 ):
     pid_b = tl.program_id(axis=1)
-    pid_ch = tl.program_id(axis=2)
+    # if chunk_size/stride products are large, may overflow int32, so use 64 bit
+    # https://github.com/triton-lang/triton/issues/1058
+    pid_ch = tl.program_id(axis=2).to(tl.int64)
     pid_c = pid_ch // ngroups
     pid_h = pid_ch - pid_c * ngroups
     num_pid_n = tl.cdiv(K, BLOCK_SIZE_N)
